refactor(accounts): extract sidebarIcon to shared file
Move the Account.sidebarIcon extension out of AccountSidebarRow so the
upcoming AccountPickerOptions view can share the same mapping. Pure
move, no behaviour change.
